[Chore] move `outline` dir under `workbench/contrib` instead of `workbench/services`
