+++ src/workbench/contrib/outline/headingItem.ts
